makes steps reusable
